backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
psycopg[binary]>=3.2.2
django-redis==5.4.0
redis==5.0.1
PyJWT==2.8.0
cryptography==42.0.2
django-encrypted-model-fields==0.6.5
python-dotenv==1.0.1
django-cors-headers==4.3.1
httpx==0.27.0
requests==2.31.0
pydantic==2.10.5
pydantic-settings==2.7.0
python-dateutil==2.8.2
pytz==2024.1
whitenoise==6.6.0
